Fix: Pre-populate database from asset
This commit enables the pre-population of the `IngredientDatabase` from a local asset file.

The `createFromAsset("ingredients_db")` method, which was previously commented out in the `AppModule.kt` dependency injection setup, has been uncommented. This ensures the Room database is initialized with data from the `ingredients_db` file located in the `assets` directory.